refactor: 추천 쿼리 서비스 분리
