+++ src/main/java/umc/cockple/demo/domain/exercise/service/query/ExerciseRecommendationQueryService.java
@@ -1,0 +1,138 @@
+package umc.cockple.demo.domain.exercise.service.query;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import umc.cockple.demo.domain.exercise.converter.ExerciseConverter;
+import umc.cockple.demo.domain.exercise.domain.Exercise;
+import umc.cockple.demo.domain.exercise.dto.ExerciseRecommendationCalendarDTO;
+import umc.cockple.demo.domain.exercise.dto.ExerciseRecommendationDTO;
+import umc.cockple.demo.domain.exercise.service.support.ExerciseBookmarkReader;
+import umc.cockple.demo.domain.exercise.service.support.ExerciseDistanceCalculator;
+import umc.cockple.demo.domain.exercise.service.support.ExerciseParticipantReader;
+import umc.cockple.demo.domain.exercise.service.support.ExerciseReader;
+import umc.cockple.demo.domain.member.domain.Member;
+import umc.cockple.demo.domain.member.domain.MemberAddr;
+import umc.cockple.demo.domain.member.service.support.MemberLookupService;
+
+import java.time.LocalDate;
+import java.util.Comparator;
+import java.util.List;
+import java.util.Map;
+
+@Service
+@Transactional(readOnly = true)
+@RequiredArgsConstructor
+@Slf4j
+public class ExerciseRecommendationQueryService {
+
+    private final ExerciseReader exerciseReader;
+    private final ExerciseBookmarkReader exerciseBookmarkReader;
+    private final ExerciseParticipantReader exerciseParticipantReader;
+    private final ExerciseDistanceCalculator exerciseDistanceCalculator;
+    private final MemberLookupService memberLookupService;
+    private final ExerciseConverter exerciseConverter;
+
+    public ExerciseRecommendationDTO.Response getRecommendedExercises(Long memberId) {
+
+        log.info("운동 추천 조회 시작 - memberId: {}", memberId);
+
+        Member member = memberLookupService.findWithAddressesOrThrow(memberId);
+        MemberAddr mainAddr = memberLookupService.findMainAddressOrThrow(member);
+
+        List<Exercise> candidateExercises = exerciseReader.findRecommendedExercises(
+                memberId, member.getGender(), member.getLevel(), member.getBirth().getYear());
+
+        List<ExerciseWithDistance> finalExercisesWithDistance = getFinalSortedExercises(candidateExercises, mainAddr);
+        List<Exercise> finalExercises = extractExercises(finalExercisesWithDistance);
+
+        List<Long> exerciseIds = getExerciseIds(finalExercises);
+        Map<Long, Boolean> bookmarkStatus = exerciseBookmarkReader.getBookmarkStatus(memberId, exerciseIds);
+
+        log.info("운동 추천 조회 종료 - memberId: {}, 결과 : {}", memberId, exerciseIds.size());
+
+        return exerciseConverter.toExerciseRecommendationResponse(finalExercises, bookmarkStatus);
+    }
+
+    public ExerciseRecommendationCalendarDTO.Response getRecommendedExerciseCalendar(
+            Long memberId,
+            LocalDate startDate,
+            LocalDate endDate,
+            Boolean isCockpleRecommend,
+            ExerciseRecommendationCalendarDTO.FilterSortType filterSortType) {
+
+        log.info("사용자 추천 운동 캘린더 조회 시작 - memberId: {}, 콕플추천: {}, 필터정렬: {}, 기간: {}~{}"
+                , memberId, isCockpleRecommend, filterSortType, startDate, endDate);
+
+        Member member = memberLookupService.findWithAddressesOrThrow(memberId);
+        DateRange dateRange = DateRange.calculateDateRange(startDate, endDate);
+
+        List<Exercise> exercises;
+
+        if (isCockpleRecommend) {
+            exercises = exerciseReader.findCockpleRecommendedByDateRange(member, dateRange.start(), dateRange.end());
+        } else {
+            exercises = exerciseReader.findFilteredRecommended(member, dateRange.start(), dateRange.end(), filterSortType);
+        }
+
+        List<Long> exerciseIds = getExerciseIds(exercises);
+        Map<Long, Boolean> bookmarkStatus = exerciseBookmarkReader.getBookmarkStatus(memberId, exerciseIds);
+        Map<Long, Integer> participantCountMap = exerciseParticipantReader.getParticipantCountsMap(exerciseIds);
+        MemberAddr mainAddr = memberLookupService.findMainAddressOrThrow(member);
+
+        log.info("사용자 추천 운동 캘린더 조회 완료 - memberId: {}, 결과 수: {}", memberId, exercises.size());
+
+        return exerciseConverter.toRecommendationCalendarResponse(
+                exercises, bookmarkStatus, participantCountMap, mainAddr
+                , dateRange.start(), dateRange.end(), isCockpleRecommend, filterSortType);
+    }
+
+    private List<ExerciseWithDistance> getFinalSortedExercises(List<Exercise> candidateExercises, MemberAddr mainAddr) {
+        return candidateExercises.stream()
+                .map(exercise -> {
+                    double distance = exerciseDistanceCalculator.calculate(
+                            mainAddr.getLatitude(),
+                            mainAddr.getLongitude(),
+                            exercise.getExerciseAddr().getLatitude(),
+                            exercise.getExerciseAddr().getLongitude()
+                    );
+                    return new ExerciseWithDistance(exercise, distance);
+                })
+                .sorted(Comparator
+                        .comparing(ExerciseWithDistance::distance)
+                        .thenComparing(ewd -> ewd.exercise().getDate())
+                        .thenComparing(ewd -> ewd.exercise().getStartTime())
+                )
+                .limit(10)
+                .toList();
+    }
+
+    private static List<Exercise> extractExercises(List<ExerciseWithDistance> finalExercisesWithDistance) {
+        return finalExercisesWithDistance.stream()
+                .map(ExerciseWithDistance::exercise)
+                .toList();
+    }
+
+    private static List<Long> getExerciseIds(List<Exercise> exercises) {
+        return exercises.stream().map(Exercise::getId).toList();
+    }
+
+    private record DateRange(LocalDate start, LocalDate end) {
+        private static DateRange calculateDateRange(LocalDate startDate, LocalDate endDate) {
+            if (startDate != null && endDate != null) {
+                return new DateRange(startDate, endDate);
+            }
+
+            LocalDate today = LocalDate.now();
+            LocalDate thisWeekMonday = today.minusDays(today.getDayOfWeek().getValue() - 1);
+            LocalDate defaultStart = thisWeekMonday.minusWeeks(1);
+            LocalDate defaultEnd = thisWeekMonday.plusWeeks(3).plusDays(6);
+
+            return new DateRange(defaultStart, defaultEnd);
+        }
+    }
+
+    private record ExerciseWithDistance(Exercise exercise, double distance) {
+    }
+}
