+++ src/test/java/umc/cockple/demo/domain/exercise/service/ExerciseQueryServiceTest.java
@@ -21,7 +21,6 @@ import umc.cockple.demo.domain.exercise.dto.ExerciseBuildingDetailDTO;
 import umc.cockple.demo.domain.exercise.dto.ExerciseDetailDTO;
 import umc.cockple.demo.domain.exercise.dto.ExerciseEditDetailDTO;
 import umc.cockple.demo.domain.exercise.dto.ExerciseMapBuildingsDTO;
-import umc.cockple.demo.domain.exercise.dto.ExerciseRecommendationCalendarDTO;
 import umc.cockple.demo.domain.exercise.dto.ExerciseMyGuestListDTO;
 import umc.cockple.demo.domain.exercise.dto.MyExerciseCalendarDTO;
 import umc.cockple.demo.domain.exercise.dto.MyExerciseListDTO;
@@ -33,11 +32,9 @@ import umc.cockple.demo.domain.exercise.enums.MyExerciseOrderType;
 import umc.cockple.demo.domain.exercise.enums.MyPartyExerciseOrderType;
 import umc.cockple.demo.domain.exercise.exception.ExerciseErrorCode;
 import umc.cockple.demo.domain.exercise.exception.ExerciseException;
-import umc.cockple.demo.domain.member.domain.MemberAddr;
 import umc.cockple.demo.domain.exercise.repository.ExerciseRepository;
 import umc.cockple.demo.domain.exercise.repository.GuestRepository;
 import umc.cockple.demo.domain.exercise.service.support.ExerciseBookmarkReader;
-import umc.cockple.demo.domain.exercise.service.support.ExerciseDistanceCalculator;
 import umc.cockple.demo.domain.exercise.service.support.ExerciseParticipantReader;
 import umc.cockple.demo.domain.exercise.service.support.ExerciseReader;
 import umc.cockple.demo.domain.exercise.service.support.GuestReader;
@@ -51,8 +48,6 @@ import umc.cockple.demo.domain.member.repository.MemberExerciseRepository;
 import umc.cockple.demo.domain.member.repository.MemberPartyRepository;
 import umc.cockple.demo.domain.member.repository.MemberRepository;
 import umc.cockple.demo.domain.party.domain.Party;
-import umc.cockple.demo.domain.party.enums.ActivityTime;
-import umc.cockple.demo.domain.party.enums.ParticipationType;
 import umc.cockple.demo.domain.party.enums.PartyStatus;
 import umc.cockple.demo.domain.party.exception.PartyErrorCode;
 import umc.cockple.demo.domain.party.exception.PartyException;
@@ -63,7 +58,6 @@ import umc.cockple.demo.global.enums.Role;
 import umc.cockple.demo.support.ExerciseCalendarTestHelper;
 import umc.cockple.demo.support.fixture.ExerciseFixture;
 import umc.cockple.demo.support.fixture.GuestFixture;
-import umc.cockple.demo.support.fixture.MemberAddrFixture;
 import umc.cockple.demo.support.fixture.MemberFixture;
 import umc.cockple.demo.support.fixture.PartyFixture;
 
@@ -132,7 +126,6 @@ class ExerciseQueryServiceTest {
                 new GuestReader(guestRepository),
                 new ExerciseParticipantReader(exerciseRepository, memberExerciseRepository, memberPartyRepository),
                 new ExerciseBookmarkReader(exerciseBookmarkRepository),
-                new ExerciseDistanceCalculator(),
                 new MemberLookupService(memberRepository),
                 new PartyLookupService(partyRepository),
                 exerciseConverter
@@ -1742,285 +1735,5 @@ class ExerciseQueryServiceTest {
         }
     }
 
-    @Nested
-    @DisplayName("getRecommendedExerciseCalendar")
-    class GetRecommendedExerciseCalendar {
 
-        private Member recommendationMember;
-        private Member memberWithoutMainAddr;
-        private MemberAddr mainAddr;
-        private Party filteredParty;
-        private LocalDate startDate;
-        private LocalDate endDate;
-
-        @BeforeEach
-        void setUp() {
-            recommendationMember = MemberFixture.createMember("추천캘린더회원", Gender.MALE, Level.A, 11001L,
-                    LocalDate.of(1995, 6, 15));
-            ReflectionTestUtils.setField(recommendationMember, "id", 11L);
-            mainAddr = MemberAddrFixture.createMainAddr(recommendationMember);
-            ReflectionTestUtils.setField(recommendationMember, "addresses", List.of(mainAddr));
-
-            memberWithoutMainAddr = MemberFixture.createMember("주소없는추천회원", Gender.MALE, Level.A, 11002L,
-                    LocalDate.of(1995, 6, 15));
-            ReflectionTestUtils.setField(memberWithoutMainAddr, "id", 12L);
-            ReflectionTestUtils.setField(memberWithoutMainAddr, "addresses", List.of(MemberAddrFixture.createSubAddr(memberWithoutMainAddr)));
-
-            party.addLevel(Gender.MALE, Level.A);
-
-            filteredParty = PartyFixture.createParty("필터 모임", manager.getId(),
-                    PartyFixture.createPartyAddr("서울특별시", "강남구"));
-            ReflectionTestUtils.setField(filteredParty, "id", 20L);
-            ReflectionTestUtils.setField(filteredParty, "partyType", ParticipationType.SINGLE);
-            ReflectionTestUtils.setField(filteredParty, "activityTime", ActivityTime.AFTERNOON);
-            filteredParty.addLevel(Gender.MALE, Level.B);
-
-            startDate = LocalDate.of(2026, 3, 23);
-            endDate = LocalDate.of(2026, 4, 5);
-        }
-
-        @Nested
-        @DisplayName("성공 케이스")
-        class Success {
-
-            @Test
-            @DisplayName("콕플 추천 기본 기간은 기본 범위를 사용하고 거리순으로 정렬한다")
-            void 콕플_추천_기본_기간은_기본_범위를_사용하고_거리순으로_정렬한다() {
-                // given
-                LocalDate expectedStart = ExerciseCalendarTestHelper.expectedDefaultStartDate();
-                LocalDate expectedEnd = ExerciseCalendarTestHelper.expectedDefaultEndDate();
-                LocalDate targetDate = expectedStart.plusDays(9);
-                int weekIndex = ExerciseCalendarTestHelper.weekIndexFor(expectedStart, targetDate);
-                int dayIndex = ExerciseCalendarTestHelper.dayIndexFor(targetDate);
-
-                Exercise nearExercise = createRecommendationExercise(party, 1001L, targetDate,
-                        LocalTime.of(11, 0), LocalTime.of(13, 0), 37.5, 127.0, "가까운 체육관");
-                Exercise farExercise = createRecommendationExercise(party, 1002L, targetDate,
-                        LocalTime.of(9, 0), LocalTime.of(11, 0), 35.1, 129.1, "먼 체육관");
-
-                given(memberRepository.findMemberWithAddresses(recommendationMember.getId()))
-                        .willReturn(Optional.of(recommendationMember));
-                given(exerciseRepository.findCockpleRecommendedExercisesByDateRange(
-                        recommendationMember.getId(), Gender.MALE, Level.A, 1995, expectedStart, expectedEnd))
-                        .willReturn(List.of(farExercise, nearExercise));
-                given(exerciseBookmarkRepository.findAllExerciseIdsByMemberIdAndExerciseIds(
-                        recommendationMember.getId(), List.of(farExercise.getId(), nearExercise.getId())))
-                        .willReturn(List.of(nearExercise.getId()));
-                given(exerciseRepository.findExerciseParticipantCountsByExerciseIds(
-                        List.of(farExercise.getId(), nearExercise.getId())))
-                        .willReturn(List.of());
-
-                // when
-                ExerciseRecommendationCalendarDTO.Response response = exerciseQueryService.getRecommendedExerciseCalendar(
-                        recommendationMember.getId(), null, null, true, recommendationFilter(MyPartyExerciseOrderType.LATEST));
-
-                // then
-                assertThat(response.startDate()).isEqualTo(expectedStart);
-                assertThat(response.endDate()).isEqualTo(expectedEnd);
-                assertThat(response.weeks()).hasSize(5);
-                assertThat(response.weeks().get(weekIndex).days().get(dayIndex).exercises())
-                        .extracting(
-                                ExerciseRecommendationCalendarDTO.ExerciseCalendarItem::exerciseId,
-                                ExerciseRecommendationCalendarDTO.ExerciseCalendarItem::partyId,
-                                ExerciseRecommendationCalendarDTO.ExerciseCalendarItem::partyName,
-                                ExerciseRecommendationCalendarDTO.ExerciseCalendarItem::buildingName,
-                                ExerciseRecommendationCalendarDTO.ExerciseCalendarItem::startTime,
-                                ExerciseRecommendationCalendarDTO.ExerciseCalendarItem::endTime,
-                                ExerciseRecommendationCalendarDTO.ExerciseCalendarItem::isBookmarked
-                        )
-                        .containsExactly(
-                                tuple(nearExercise.getId(), party.getId(), "테스트 모임", "가까운 체육관",
-                                        LocalTime.of(11, 0), LocalTime.of(13, 0), true),
-                                tuple(farExercise.getId(), party.getId(), "테스트 모임", "먼 체육관",
-                                        LocalTime.of(9, 0), LocalTime.of(11, 0), false)
-                        );
-                assertThat(response.weeks().get(weekIndex).days().get(dayIndex).exercises().get(0).distance()).isZero();
-                assertThat(response.weeks().get(weekIndex).days().get(dayIndex).exercises().get(1).distance()).isGreaterThan(0.0);
-                verify(exerciseRepository).findCockpleRecommendedExercisesByDateRange(
-                        recommendationMember.getId(), Gender.MALE, Level.A, 1995, expectedStart, expectedEnd);
-                verify(exerciseRepository, never()).findFilteredRecommendedExercisesForCalendar(any(), any(), any(), any(), any());
-            }
-
-            @Test
-            @DisplayName("필터 추천은 필터 리포지토리만 호출하고 인기순 정렬을 적용한다")
-            void 필터_추천은_필터_리포지토리만_호출하고_인기순_정렬을_적용한다() {
-                // given
-                Exercise popularExercise = createRecommendationExercise(filteredParty, 1101L, LocalDate.of(2026, 3, 25),
-                        LocalTime.of(18, 0), LocalTime.of(20, 0), 37.52, 127.02, "인기 체육관");
-                Exercise earlyExercise = createRecommendationExercise(filteredParty, 1102L, LocalDate.of(2026, 3, 25),
-                        LocalTime.of(9, 0), LocalTime.of(11, 0), 37.53, 127.03, "이른 체육관");
-
-                ExerciseRecommendationCalendarDTO.FilterSortType filterSortType = ExerciseRecommendationCalendarDTO.FilterSortType.builder()
-                        .addr1("서울특별시")
-                        .addr2("강남구")
-                        .levels(List.of(Level.B))
-                        .participationTypes(List.of(ParticipationType.SINGLE))
-                        .activityTimes(List.of(ActivityTime.AFTERNOON))
-                        .sortType(MyPartyExerciseOrderType.POPULARITY)
-                        .build();
-
-                given(memberRepository.findMemberWithAddresses(recommendationMember.getId()))
-                        .willReturn(Optional.of(recommendationMember));
-                given(exerciseRepository.findFilteredRecommendedExercisesForCalendar(
-                        recommendationMember.getId(), 1995, filterSortType, startDate, endDate))
-                        .willReturn(List.of(earlyExercise, popularExercise));
-                given(exerciseBookmarkRepository.findAllExerciseIdsByMemberIdAndExerciseIds(
-                        recommendationMember.getId(), List.of(earlyExercise.getId(), popularExercise.getId())))
-                        .willReturn(List.of(popularExercise.getId()));
-                given(exerciseRepository.findExerciseParticipantCountsByExerciseIds(
-                        List.of(earlyExercise.getId(), popularExercise.getId())))
-                        .willReturn(List.of(
-                                new Object[]{popularExercise.getId(), 3},
-                                new Object[]{earlyExercise.getId(), 1}
-                        ));
-
-                // when
-                ExerciseRecommendationCalendarDTO.Response response = exerciseQueryService.getRecommendedExerciseCalendar(
-                        recommendationMember.getId(), startDate, endDate, false, filterSortType);
-
-                // then
-                assertThat(response.startDate()).isEqualTo(startDate);
-                assertThat(response.endDate()).isEqualTo(endDate);
-                assertThat(response.weeks().get(0).days().get(2).exercises())
-                        .extracting(
-                                ExerciseRecommendationCalendarDTO.ExerciseCalendarItem::exerciseId,
-                                ExerciseRecommendationCalendarDTO.ExerciseCalendarItem::partyId,
-                                ExerciseRecommendationCalendarDTO.ExerciseCalendarItem::partyName,
-                                ExerciseRecommendationCalendarDTO.ExerciseCalendarItem::buildingName,
-                                ExerciseRecommendationCalendarDTO.ExerciseCalendarItem::isBookmarked,
-                                ExerciseRecommendationCalendarDTO.ExerciseCalendarItem::distance
-                        )
-                        .containsExactly(
-                                tuple(popularExercise.getId(), filteredParty.getId(), "필터 모임", "인기 체육관", true, null),
-                                tuple(earlyExercise.getId(), filteredParty.getId(), "필터 모임", "이른 체육관", false, null)
-                        );
-                verify(exerciseRepository, never()).findCockpleRecommendedExercisesByDateRange(any(), any(), any(), anyInt(), any(), any());
-                verify(exerciseRepository).findFilteredRecommendedExercisesForCalendar(
-                        recommendationMember.getId(), 1995, filterSortType, startDate, endDate);
-            }
-
-            @Test
-            @DisplayName("추천 운동이 없으면 기간 메타데이터와 빈 일자별 캘린더를 반환한다")
-            void 추천_운동이_없으면_기간_메타데이터와_빈_일자별_캘린더를_반환한다() {
-                // given
-                given(memberRepository.findMemberWithAddresses(recommendationMember.getId()))
-                        .willReturn(Optional.of(recommendationMember));
-                given(exerciseRepository.findCockpleRecommendedExercisesByDateRange(
-                        recommendationMember.getId(), Gender.MALE, Level.A, 1995, startDate, endDate))
-                        .willReturn(List.of());
-
-                // when
-                ExerciseRecommendationCalendarDTO.Response response = exerciseQueryService.getRecommendedExerciseCalendar(
-                        recommendationMember.getId(), startDate, endDate, true, recommendationFilter(MyPartyExerciseOrderType.LATEST));
-
-                // then
-                assertThat(response.startDate()).isEqualTo(startDate);
-                assertThat(response.endDate()).isEqualTo(endDate);
-                assertThat(response.weeks()).hasSize(2);
-                assertThat(response.weeks().get(0).days()).hasSize(7);
-                assertThat(response.weeks().get(0).days().get(0).exercises()).isEmpty();
-            }
-
-            @Test
-            @DisplayName("startDate만 주어져도 기본 기간이 적용된다")
-            void startDate만_주어져도_기본_기간이_적용된다() {
-                // given
-                LocalDate expectedStart = ExerciseCalendarTestHelper.expectedDefaultStartDate();
-                LocalDate expectedEnd = ExerciseCalendarTestHelper.expectedDefaultEndDate();
-
-                given(memberRepository.findMemberWithAddresses(recommendationMember.getId()))
-                        .willReturn(Optional.of(recommendationMember));
-                given(exerciseRepository.findCockpleRecommendedExercisesByDateRange(
-                        recommendationMember.getId(), Gender.MALE, Level.A, 1995, expectedStart, expectedEnd))
-                        .willReturn(List.of());
-
-                // when
-                ExerciseRecommendationCalendarDTO.Response response = exerciseQueryService.getRecommendedExerciseCalendar(
-                        recommendationMember.getId(), LocalDate.of(2026, 3, 25), null, true,
-                        recommendationFilter(MyPartyExerciseOrderType.LATEST));
-
-                // then
-                assertThat(response.startDate()).isEqualTo(expectedStart);
-                assertThat(response.endDate()).isEqualTo(expectedEnd);
-                assertThat(response.weeks()).hasSize(5);
-            }
-
-            @Test
-            @DisplayName("종료일이 시작일보다 이전이어도 빈 캘린더를 반환한다")
-            void 종료일이_시작일보다_이전이어도_빈_캘린더를_반환한다() {
-                // given
-                LocalDate reversedStart = LocalDate.of(2026, 4, 5);
-                LocalDate reversedEnd = LocalDate.of(2026, 3, 23);
-
-                given(memberRepository.findMemberWithAddresses(recommendationMember.getId()))
-                        .willReturn(Optional.of(recommendationMember));
-                given(exerciseRepository.findCockpleRecommendedExercisesByDateRange(
-                        recommendationMember.getId(), Gender.MALE, Level.A, 1995, reversedStart, reversedEnd))
-                        .willReturn(List.of());
-
-                // when
-                ExerciseRecommendationCalendarDTO.Response response = exerciseQueryService.getRecommendedExerciseCalendar(
-                        recommendationMember.getId(), reversedStart, reversedEnd, true,
-                        recommendationFilter(MyPartyExerciseOrderType.LATEST));
-
-                // then
-                assertThat(response.startDate()).isEqualTo(reversedStart);
-                assertThat(response.endDate()).isEqualTo(reversedEnd);
-                assertThat(response.weeks()).isEmpty();
-            }
-        }
-
-        @Nested
-        @DisplayName("실패 케이스")
-        class Failure {
-
-            @Test
-            @DisplayName("존재하지 않는 멤버면 예외를 던진다")
-            void 존재하지_않는_멤버면_예외를_던진다() {
-                // given
-                given(memberRepository.findMemberWithAddresses(999L))
-                        .willReturn(Optional.empty());
-
-                // when & then
-                assertThatThrownBy(() -> exerciseQueryService.getRecommendedExerciseCalendar(
-                        999L, startDate, endDate, true, recommendationFilter(MyPartyExerciseOrderType.LATEST)))
-                        .isInstanceOf(ExerciseException.class)
-                        .hasFieldOrPropertyWithValue("code", ExerciseErrorCode.MEMBER_NOT_FOUND);
-            }
-
-            @Test
-            @DisplayName("대표주소가 없으면 예외를 던진다")
-            void 대표주소가_없으면_예외를_던진다() {
-                // given
-                given(memberRepository.findMemberWithAddresses(memberWithoutMainAddr.getId()))
-                        .willReturn(Optional.of(memberWithoutMainAddr));
-                given(exerciseRepository.findCockpleRecommendedExercisesByDateRange(
-                        memberWithoutMainAddr.getId(), Gender.MALE, Level.A, 1995, startDate, endDate))
-                        .willReturn(List.of());
-
-                // when & then
-                assertThatThrownBy(() -> exerciseQueryService.getRecommendedExerciseCalendar(
-                        memberWithoutMainAddr.getId(), startDate, endDate, true, recommendationFilter(MyPartyExerciseOrderType.LATEST)))
-                        .isInstanceOf(ExerciseException.class)
-                        .hasFieldOrPropertyWithValue("code", ExerciseErrorCode.MAIN_ADDRESS_NULL);
-            }
-        }
-
-        private ExerciseRecommendationCalendarDTO.FilterSortType recommendationFilter(MyPartyExerciseOrderType sortType) {
-            return ExerciseRecommendationCalendarDTO.FilterSortType.builder()
-                    .sortType(sortType)
-                    .build();
-        }
-
-        private Exercise createRecommendationExercise(Party exerciseParty, long id, LocalDate date,
-                                                      LocalTime startTime, LocalTime endTime,
-                                                      double latitude, double longitude, String buildingName) {
-            Exercise recommendationExercise = ExerciseFixture.createRecommendableExercise(
-                    exerciseParty, date, latitude, longitude, buildingName);
-            ReflectionTestUtils.setField(recommendationExercise, "id", id);
-            ReflectionTestUtils.setField(recommendationExercise, "startTime", startTime);
-            ReflectionTestUtils.setField(recommendationExercise, "endTime", endTime);
-            return recommendationExercise;
-        }
-    }
 }
